clovers_runtime::trace

Function trace

pub fn trace(
    ray: &Ray,
    scene: &Scene<'_>,
    depth: u32,
    max_depth: u32,
    rng: &mut SmallRng,
    sampler: &dyn SamplerTrait<'_>,
) -> [Float; 4]

The main path tracing function. Sends a Ray to the Scene, sees if it hits anything, and eventually returns a spectral intensity. Taking into account the Material that is hit, the method recurses with various adjustments, with a new Ray started from the location that was hit.